gh issue #3: content addressable manifests (part 1)
* digest must include config from metadata in its hash, and it should or must(?) include all the layers of the manifest
We will need to make sure to have unique digests, therefore this commit adds cname as a step towards uniquene oci configs.

* todo: add version to digest hash calc

* todo: add layers to digest hash calc.
This will need special case treatment for incrementally added layers such as test logs.
Will the digest change, and we overwrite the previous version? Need to clarify before continue with part 2 of this change