digest calculation
Manifest has unique tag, use unique tag to download manifest and then create hash for response content from registry
the response from the registry SHOULD exactly be the same content used by the registry to calculate the digest.

This way, we do not have different order of attributes, different whitespaces or other stuff that changes the hash.

Currently, I do not find another way to make sure that the hash used to reference the manifest in the index is compliant to
what the registry calculated.