gh issue #15: calculate size
oci-spec requires manifest entries in image-index to be a descriptor.
a descriptor must have a size attribute.

Clients SHOULD reject artifacts that have a missmatch in size between downloaded data and size attribute given in descriptor.
Therefore, we must calculate size according to oci specification.

It is specified in oci-spec, that the size is in bytes of raw content. It is not further specified if this includes the
json manifest, or only the sum of all layers.

This commit ASSUMES that it is the sum of all layers in size, excluding manifest json.